+++ lib/rubocop/cop/layout/indent_array.rb
@@ -82,7 +82,7 @@ module RuboCop
       class IndentArray < Cop
         include Alignment
         include ConfigurableEnforcedStyle
-        include ArrayHashIndentation
+        include MultilineElementIndentation
 
         MSG = 'Use %<configured_indentation_width>d spaces for indentation ' \
               'in an array, relative to %<base_description>s.'.freeze